+++ tests/testthat/test-elemental_conversions.R
@@ -39,6 +39,42 @@ test_that("current_LCU_2_constant_LCU_base_x wb_wdi", {
   }
 })
 
+# test_that("current_LCU_2_constant_LCU_base_y linked series", {
+#   country_base_years <- wb_wdi_linked %>%
+#     dplyr::filter(`GDP deflator` == 1) %>%
+#     dplyr::select("iso3c", "year")
+#
+#   my_years <- country_base_years %>%
+#     dplyr::group_by(year) %>%
+#     dplyr::count() %>%
+#     dplyr::arrange(-n) %>%
+#     dplyr::pull(year)
+#
+#   for (my_base_year in my_years) {
+#     my_countries <- country_base_years %>%
+#       dplyr::filter(year == my_base_year) %>%
+#       dplyr::pull(iso3c)
+#
+#     gdp_in <- wb_wdi %>%
+#       dplyr::filter(iso3c %in% my_countries,
+#                     !is.na(`GDP (constant LCU)`)) %>%
+#       dplyr::select("iso3c", "year", "value" = `GDP: linked series (current LCU)`)
+#
+#     gdp_conv <- current_LCU_2_constant_LCU_base_y(gdp_in, my_base_year, wb_wdi_linked) %>%
+#       dplyr::filter(!is.na(value))
+#
+#     gdp_out <- wb_wdi %>%
+#       dplyr::right_join(gdp_conv, by = c("iso3c", "year")) %>%
+#       dplyr::select("iso3c", "year", "value" = `GDP (constant LCU)`)
+#
+#     for (country in unique(gdp_conv$iso3c)) {
+#       expect_equal(gdp_conv %>% dplyr::filter(iso3c == country),
+#                    gdp_out %>% dplyr::filter(iso3c == country),
+#                    label = country)
+#     }
+#   }
+# })
+
 test_that("current_LCU_2_current_IntPPP", {
   gdp_in <- wb_wdi %>%
     dplyr::filter(!is.na(`GDP, PPP (current international $)`)) %>%
@@ -143,7 +179,7 @@ test_that("constant_LCU_2_constant_IntPPP", {
 
   gdp_in <- wb_wdi %>%
     dplyr::filter(iso3c %in% my_countries,
-                  !is.na(dplyr::across(tidyselect::matches(regex_var_IntPPP)))) %>%
+                  dplyr::if_all(tidyselect::matches(regex_var_IntPPP), ~!is.na(.x))) %>%
     dplyr::select("iso3c", "year", "value" = `GDP (constant LCU)`)
 
   gdp_conv <- constant_LCU_2_constant_IntPPP(gdp_in, year_IntPPP, wb_wdi) %>%
@@ -171,7 +207,7 @@ test_that("constant_LCU_2_constant_USMER", {
 
   gdp_in <- wb_wdi %>%
     dplyr::filter(iso3c %in% my_countries,
-                  !is.na(dplyr::across(tidyselect::matches(regex_var_USMER)))) %>%
+                  dplyr::if_all(tidyselect::matches(regex_var_USMER), ~!is.na(.x))) %>%
     dplyr::select("iso3c", "year", "value" = `GDP (constant LCU)`)
 
   gdp_conv <- constant_LCU_2_constant_USMER(gdp_in, year_USMER, wb_wdi) %>%